requirements.txt
requests>=2.25.0
python-dotenv>=0.21.0
openai>=0.27.0
selenium>=4.0.0
webdriver-manager>=3.8.6
